requirements.txt
astroid==2.0.4
autopep8==1.3.5
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1
django-rest-swagger==2.2.0
djangorestframework==3.8.2
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
openapi-codec==1.3.2
pycodestyle==2.4.0
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
six==1.11.0
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.23
wrapt==1.10.11
